Safari-611.1.10/LayoutTests/media/video-add-autoplay-user-gesture.html

<!DOCTYPE html>
<html>
<head>
    <title>video-add-autoplay-user-gesture</title>
    <script src=media-file.js></script>
    <script src=video-test.js></script>
    <script>
    function runTest() {
        run('window.internals.settings.setVideoPlaybackRequiresUserGesture(true);');
        run('video = document.createElement("video")');
        run('video.src = findMediaFile("video", "content/test")');
        runWithKeyDown(() => {
            run('video.setAttribute("autoplay", "")');
            run('document.body.appendChild(video)');
            waitForEventAndEnd('playing');
        });
    }
    </script>
</head>
<body onload="runTest()">
</body>
</html>
